requirements.txt
alabaster==0.7.12
applaunchservices==0.2.1
appnope==0.1.2
arrow==1.2.2
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cookiecutter==1.7.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
docutils==0.17.1
entrypoints==0.4
filelock==3.6.0
flake8==4.0.1
huggingface-hub==0.4.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.2.0
importlib-resources==5.4.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
keyring==23.5.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.21.5
numpydoc==1.2
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.12.2
pyls-spyder==0.4.0
pyobjc==8.3
pyobjc-core==8.3
pyobjc-framework-Accessibility==8.3
pyobjc-framework-Accounts==8.3
pyobjc-framework-AddressBook==8.3
pyobjc-framework-AdServices==8.3
pyobjc-framework-AdSupport==8.3
pyobjc-framework-AppleScriptKit==8.3
pyobjc-framework-AppleScriptObjC==8.3
pyobjc-framework-ApplicationServices==8.3
pyobjc-framework-AppTrackingTransparency==8.3
pyobjc-framework-AudioVideoBridging==8.3
pyobjc-framework-AuthenticationServices==8.3
pyobjc-framework-AutomaticAssessmentConfiguration==8.3
pyobjc-framework-Automator==8.3
pyobjc-framework-AVFoundation==8.3
pyobjc-framework-AVKit==8.3
pyobjc-framework-BusinessChat==8.3
pyobjc-framework-CalendarStore==8.3
pyobjc-framework-CallKit==8.3
pyobjc-framework-CFNetwork==8.3
pyobjc-framework-ClassKit==8.3
pyobjc-framework-CloudKit==8.3
pyobjc-framework-Cocoa==8.3
pyobjc-framework-Collaboration==8.3
pyobjc-framework-ColorSync==8.3
pyobjc-framework-Contacts==8.3
pyobjc-framework-ContactsUI==8.3
pyobjc-framework-CoreAudio==8.3
pyobjc-framework-CoreAudioKit==8.3
pyobjc-framework-CoreBluetooth==8.3
pyobjc-framework-CoreData==8.3
pyobjc-framework-CoreHaptics==8.3
pyobjc-framework-CoreLocation==8.3
pyobjc-framework-CoreMedia==8.3
pyobjc-framework-CoreMediaIO==8.3
pyobjc-framework-CoreMIDI==8.3
pyobjc-framework-CoreML==8.3
pyobjc-framework-CoreMotion==8.3
pyobjc-framework-CoreServices==8.3
pyobjc-framework-CoreSpotlight==8.3
pyobjc-framework-CoreText==8.3
pyobjc-framework-CoreWLAN==8.3
pyobjc-framework-CryptoTokenKit==8.3
pyobjc-framework-DataDetection==8.3
pyobjc-framework-DeviceCheck==8.3
pyobjc-framework-DictionaryServices==8.3
pyobjc-framework-DiscRecording==8.3
pyobjc-framework-DiscRecordingUI==8.3
pyobjc-framework-DiskArbitration==8.3
pyobjc-framework-DVDPlayback==8.3
pyobjc-framework-EventKit==8.3
pyobjc-framework-ExceptionHandling==8.3
pyobjc-framework-ExecutionPolicy==8.3
pyobjc-framework-ExternalAccessory==8.3
pyobjc-framework-FileProvider==8.3
pyobjc-framework-FileProviderUI==8.3
pyobjc-framework-FinderSync==8.3
pyobjc-framework-FSEvents==8.3
pyobjc-framework-GameCenter==8.3
pyobjc-framework-GameController==8.3
pyobjc-framework-GameKit==8.3
pyobjc-framework-GameplayKit==8.3
pyobjc-framework-ImageCaptureCore==8.3
pyobjc-framework-IMServicePlugIn==8.3
pyobjc-framework-InputMethodKit==8.3
pyobjc-framework-InstallerPlugins==8.3
pyobjc-framework-InstantMessage==8.3
pyobjc-framework-Intents==8.3
pyobjc-framework-IntentsUI==8.3
pyobjc-framework-IOSurface==8.3
pyobjc-framework-iTunesLibrary==8.3
pyobjc-framework-KernelManagement==8.3
pyobjc-framework-LatentSemanticMapping==8.3
pyobjc-framework-LaunchServices==8.3
pyobjc-framework-libdispatch==8.3
pyobjc-framework-LinkPresentation==8.3
pyobjc-framework-LocalAuthentication==8.3
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.3
pyobjc-framework-MailKit==8.3
pyobjc-framework-MapKit==8.3
pyobjc-framework-MediaAccessibility==8.3
pyobjc-framework-MediaLibrary==8.3
pyobjc-framework-MediaPlayer==8.3
pyobjc-framework-MediaToolbox==8.3
pyobjc-framework-Metal==8.3
pyobjc-framework-MetalKit==8.3
pyobjc-framework-MetalPerformanceShaders==8.3
pyobjc-framework-MetalPerformanceShadersGraph==8.3
pyobjc-framework-MetricKit==8.3
pyobjc-framework-MLCompute==8.3
pyobjc-framework-ModelIO==8.3
pyobjc-framework-MultipeerConnectivity==8.3
pyobjc-framework-NaturalLanguage==8.3
pyobjc-framework-NetFS==8.3
pyobjc-framework-Network==8.3
pyobjc-framework-NetworkExtension==8.3
pyobjc-framework-NotificationCenter==8.3
pyobjc-framework-OpenDirectory==8.3
pyobjc-framework-OSAKit==8.3
pyobjc-framework-OSLog==8.3
pyobjc-framework-PassKit==8.3
pyobjc-framework-PencilKit==8.3
pyobjc-framework-Photos==8.3
pyobjc-framework-PhotosUI==8.3
pyobjc-framework-PreferencePanes==8.3
pyobjc-framework-PushKit==8.3
pyobjc-framework-Quartz==8.3
pyobjc-framework-QuickLookThumbnailing==8.3
pyobjc-framework-ReplayKit==8.3
pyobjc-framework-SafariServices==8.3
pyobjc-framework-SceneKit==8.3
pyobjc-framework-ScreenSaver==8.3
pyobjc-framework-ScreenTime==8.3
pyobjc-framework-ScriptingBridge==8.3
pyobjc-framework-SearchKit==8.3
pyobjc-framework-Security==8.3
pyobjc-framework-SecurityFoundation==8.3
pyobjc-framework-SecurityInterface==8.3
pyobjc-framework-ServiceManagement==8.3
pyobjc-framework-ShazamKit==8.3
pyobjc-framework-Social==8.3
pyobjc-framework-SoundAnalysis==8.3
pyobjc-framework-Speech==8.3
pyobjc-framework-SpriteKit==8.3
pyobjc-framework-StoreKit==8.3
pyobjc-framework-SyncServices==8.3
pyobjc-framework-SystemConfiguration==8.3
pyobjc-framework-SystemExtensions==8.3
pyobjc-framework-UniformTypeIdentifiers==8.3
pyobjc-framework-UserNotifications==8.3
pyobjc-framework-UserNotificationsUI==8.3
pyobjc-framework-VideoSubscriberAccount==8.3
pyobjc-framework-VideoToolbox==8.3
pyobjc-framework-Virtualization==8.3
pyobjc-framework-Vision==8.3
pyobjc-framework-WebKit==8.3
pyparsing==3.0.7
PyQt5==5.12.3
PyQt5-sip==12.9.1
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-lsp-black==1.1.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.1
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.2
requests==2.27.1
rope==0.23.0
Rtree==0.9.7
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.2.2
spyder-kernels==2.2.1
testpath==0.6.0
text-unidecode==1.3
textdistance==4.2.2
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss2==1.1.1
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
torch==1.10.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.16.2
typed-ast==1.5.2
typing_extensions==4.1.1
ujson==5.1.0
urllib3==1.26.8
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
wurlitzer==3.0.2
yapf==0.32.0
zipp==3.7.0
